Add files via upload
This file contains the results of part B of the round robin study on the determination of weld geometry parameters involving 22 participants. It includes detailed data for all three specimens and the corresponding weld toes, as well as device specifications. The dataset and findings are structured to facilitate analysis and reproducibility for researchers and developers interested in optical scanning technology for welded joints.